pyproject.toml
[tool.poetry]
name = "torrent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.1"
qbittorrent-api = "^2025.5.0"
beautifulsoup4 = "^4.13.4"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
